farmease-app/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
httpx==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
tensorflow==2.16.1
scikit-learn==1.4.1
joblib==1.3.2
